+++ package.json
@@ -18,6 +18,7 @@
     "ejs": "^3.1.10",
     "express": "^4.21.0",
     "pg": "^8.13.0",
+    "reflect-metadata": "^0.2.2",
     "typeorm": "^0.3.20"
   },
   "devDependencies": {
