+++ src/daos/IDAO.ts
@@ -1,0 +1,3 @@
+export default interface IDAO<T> {
+    salvar(entidade: T): Promise<T>;
+}
