+++ src/entities/Cliente.ts
@@ -1,86 +1,89 @@
-import { Entity, Column, OneToMany } from "typeorm";
+import {Entity, Column, OneToMany, OneToOne, JoinColumn} from "typeorm";
 import Endereco from "./Endereco";
 import Telefone from "./Telefone";
 import Cartao from "./Cartao";
 import Usuario from "./Usuario";
 import Genero from "../enums/Genero";
+import Entidade from "./Entidade";
 
-@Entity("tb_usuario")
-export default class Cliente extends Usuario {
-  @Column({ type: "enum", enum: Genero })
-  genero!: Genero;
+@Entity("tb_cliente")
+export default class Cliente extends Entidade {
+    @Column({type: "enum", enum: Genero})
+    genero!: Genero;
 
-  @Column({ type: "date" })
-  dtNascimento!: Date;
+    @Column({type: "date"})
+    dtNascimento!: Date;
 
-  @Column({ type: "varchar" })
-  cpf!: string;
+    @Column({type: "varchar"})
+    cpf!: string;
+    
+    @OneToMany(
+        () => Telefone,
+        (telefone: Telefone) => telefone.cliente,
+        {cascade: true, eager: true}
+    )
+    telefones!: Telefone[];
 
-  @OneToMany(
-      () => Telefone,
-      (telefone: Telefone) => telefone.cliente,
-      { cascade: true, eager: true }
-  )
-  telefones: Telefone[] = [];
+    @OneToMany(
+        () => Cartao,
+        (cartao: Cartao) => cartao.cliente,
+        {cascade: true, eager: true}
+    )
+    cartoes!: Cartao[];
 
-  @OneToMany(
-      () => Cartao,
-      (cartao: Cartao) => cartao.cliente,
-      { cascade: true, eager: true }
-  )
-  cartoes: Cartao[] = [];
+    @OneToMany(
+        () => Endereco,
+        (endereco: Endereco) => endereco.cliente,
+        {cascade: true, eager: true}
+    )
+    enderecos!: Endereco[];
 
-  @OneToMany(
-      () => Endereco,
-      (endereco: Endereco) => endereco.cliente,
-      { cascade: true, eager: true }
-  )
-  enderecos: Endereco[] = [];
+    @OneToOne(() => Usuario, {cascade: true})
+    @JoinColumn()
+    usuario!: Usuario;
 
-  @Column({ type: "int" })
-  ranking!: number;
+    @Column({type: "int", nullable: true})
+    ranking?: number;
 
-  constructor(
-    genero: Genero,
-    nome: string,
-    dtNascimento: Date,
-    cpf: string,
-    telefones: Telefone[],
-    cartoes: Cartao[],
-    enderecos: Endereco[],
-    email: string,
-    senha: string
-  ) {
-    super(email, senha, nome);
-    this.genero = genero;
-    this.nome = nome;
-    this.dtNascimento = dtNascimento;
-    this.cpf = cpf;
-    this.telefones = telefones;
-    this.cartoes = cartoes;
-    this.enderecos = enderecos;
-  }
-
-  validarDadosObrigatorios(): void {
-    this.validarCpf();
-
-    for (const telefone of this.telefones) {
-      telefone.validarDadosTelefone();
+    constructor(
+        genero: Genero,
+        dtNascimento: Date,
+        cpf: string,
+        telefones: Telefone[],
+        cartoes: Cartao[],
+        enderecos: Endereco[],
+        usuario: Usuario
+    ) {
+        super();
+        this.genero = genero;
+        this.dtNascimento = dtNascimento;
+        this.cpf = cpf;
+        this.telefones = telefones;
+        this.cartoes = cartoes;
+        this.enderecos = enderecos;
+        this.usuario = usuario;
     }
 
-    for (const cartao of this.cartoes) {
-      cartao.validarDadosCartao();
+    validarDadosObrigatorios(): void {
+        this.validarCpf();
+
+        for (const telefone of this.telefones) {
+            telefone.validarDadosTelefone();
+        }
+
+        for (const cartao of this.cartoes) {
+            cartao.validarDadosCartao();
+        }
+
+        for (const endereco of this.enderecos) {
+            endereco.validarDadosEndereco();
+        }
     }
 
-    for (const endereco of this.enderecos) {
-      endereco.validarDadosEndereco();
+    private validarCpf(): void {
+        const cpfRegex: RegExp = /^\d{3}\.\d{3}\.\d{3}-\d{2}$/
+        if (!cpfRegex.test(this.cpf)) {
+            throw new Error(`CPF inválido: ${this.cpf}`);
+        }
     }
-  }
-
-  private validarCpf(): void {
-    const cpfRegex: RegExp = /^\d{3}\.\d{3}\.\d{3}-\d{2}$/
-    if (!cpfRegex.test(this.cpf)) {
-      throw new Error(`CPF inválido: ${this.cpf}`);
-    }
-  }
 }
