--- src/logs/criacaoDeCliente.txt
@@ -1,1 +1,0 @@
-O usuário Juliana foi criado com sucesso em Tue Oct 15 2024 08:41:11 GMT-0300 (Horário Padrão de Brasília)
