+++ src/server.ts
@@ -1,11 +1,12 @@
 import "reflect-metadata";
 import app from "./app";
-import { initializeDatabase } from "./config/database/dataSource";
+import ConexaoTypeORM from "./config/database/ConexaoTypeORM";
+import { postgresDataSource } from "./config/database/dataSources/postgresDataSource";
 
 const PORTA = 3000;
 
 async function startServer() {
-  await initializeDatabase();
+  await ConexaoTypeORM.conectar(postgresDataSource);
 
   app.listen(PORTA, () => {
     console.log(`Servidor rodando na porta http://localhost:${PORTA}`);
